rename enumerator because of error "cannot load such file -- enumerator"
